tests/requirements.txt
pyyaml>=6.0
click>=8.0
pytest>=7.0
pytest-cov>=4.0
